Modify vagrant file and scripts

+++ Vagrantfile
@@ -2,24 +2,25 @@ Vagrant.configure("2") do |config|
   config.vm.box = "ubuntu/xenial64"
 
   config.vm.define :tbl do |web_config|
-    config.vm.network "forwarded_port", guest: 8000, host: 8080
+    config.vm.network "forwarded_port", guest: 80, host: 8000
     config.vm.provision "shell", inline: <<-SHELL
       # Update apt-get
       apt-get update
 
       # Install dependencies
       apt-get install -y \
-      apt-transport-https \
-      ca-certificates \
-      curl \
-      software-properties-common \
-      nginx vim git
+        apt-transport-https \
+        ca-certificates \
+        curl \
+        software-properties-common \
+        nginx vim git \
+        python3-dev \
+        python3-pip \
+        libpq-dev \
+        build-essential
 
       # Modify language to PT-BR
       locale-gen pt_BR.UTF-8
-
-      # Clone the TBL repository
-      git clone https://github.com/VictorArnaud/TBL.git
 
       # Get the GPG docker key
       curl -fsSL https://download.docker.com/linux/debian/gpg | apt-key add - && apt-key fingerprint 0EBFCD88
@@ -32,26 +33,36 @@ Vagrant.configure("2") do |config|
 
       docker --version && docker-compose --version
 
+      # Clone the TBL repository
+      if [ ! -d "/home/vagrant/TBL" ]; then
+        git clone https://github.com/VictorArnaud/TBL.git
+      fi
+      cd /home/vagrant/TBL
+      pip3 install --upgrade pip
+      pip3 install -r pgtbl/requirements.txt
+
       # Run deploy enviroment
-      cd TBL
-      mv continuos_deploy.sh.example continuos_deploy.sh
+      if [ ! -f "continuos_deploy.sh" ]; then
+        mv continuos_deploy.sh.example continuos_deploy.sh
+      fi
       chmod +x continuos_deploy.sh
       docker-compose -f docker-compose.deploy.yml up -d --build
 
       # Config NGINX
       # Copying the nginx configuration file into the container
-      cp ./images/nginx/nginx.conf /etc/nginx/conf.d/nginx.conf
+      cp nginx.conf /etc/nginx/conf.d/nginx.conf
       # Removing the nginx default page
       rm -rf /usr/share/nginx/html/*
+      rm -rf /etc/nginx/sites-enabled/* && rm -rf /etc/nginx/sites-available/*
 
       # Pick up the static files and insert them inside the nginx repository so that they are served
+	    python3 pgtbl/manage.py collectstatic --noinput
       cp ./pgtbl/tbl/staticfiles/ /usr/share/nginx/html
       cp ./pgtbl/tbl/mediafiles/ /usr/share/nginx/html
 
       # Run nginx
       service nginx start
+      systemctl status nginx.service
     SHELL
   end
-
-
 end
